+++ .gitignore
@@ -31,6 +31,7 @@ yarn-error.log*
 
 # turbo
 .turbo
+.turbo-cookie
 
 # vercel
 .vercel
