+++ apps/marketing/package.json
@@ -21,7 +21,7 @@
     "framer-motion": "^10.12.8",
     "lucide-react": "^0.279.0",
     "micro": "^10.0.1",
-    "next": "13.4.19",
+    "next": "13.5.4",
     "next-auth": "4.22.3",
     "next-contentlayer": "^0.3.4",
     "next-plausible": "^3.10.1",
@@ -34,7 +34,7 @@
     "react-icons": "^4.11.0",
     "recharts": "^2.7.2",
     "sharp": "0.32.5",
-    "typescript": "5.1.6",
+    "typescript": "5.2.2",
     "zod": "^3.21.4"
   },
   "devDependencies": {
