+++ apps/web/package.json
@@ -25,7 +25,7 @@
     "lucide-react": "^0.279.0",
     "luxon": "^3.4.0",
     "micro": "^10.0.1",
-    "next": "13.4.19",
+    "next": "13.5.4",
     "next-auth": "4.22.3",
     "next-plausible": "^3.10.1",
     "next-themes": "^0.2.1",
@@ -40,7 +40,7 @@
     "react-rnd": "^10.4.1",
     "sharp": "0.32.5",
     "ts-pattern": "^5.0.5",
-    "typescript": "5.1.6",
+    "typescript": "5.2.2",
     "zod": "^3.21.4"
   },
   "devDependencies": {
