+++ packages/ee/package.json
@@ -17,7 +17,7 @@
     "@documenso/prisma": "*",
     "luxon": "^3.4.0",
     "micro": "^10.0.1",
-    "next": "13.4.19",
+    "next": "13.5.4",
     "next-auth": "4.22.3",
     "react": "18.2.0",
     "ts-pattern": "^5.0.5",
