+++ packages/email/render.ts
@@ -1,1 +1,1 @@
-export { render, renderAsync } from '@react-email/components';
+export { render } from '@react-email/components';
