+++ packages/eslint-config/package.json
@@ -7,8 +7,8 @@
     "clean": "rimraf node_modules"
   },
   "dependencies": {
-    "@typescript-eslint/eslint-plugin": "^5.59.2",
-    "@typescript-eslint/parser": "^5.59.2",
+    "@typescript-eslint/eslint-plugin": "6.8.0",
+    "@typescript-eslint/parser": "6.8.0",
     "eslint": "^8.40.0",
     "eslint-config-next": "13.4.19",
     "eslint-config-prettier": "^8.8.0",
@@ -16,6 +16,6 @@
     "eslint-plugin-package-json": "^0.1.4",
     "eslint-plugin-prettier": "^4.2.1",
     "eslint-plugin-react": "^7.32.2",
-    "typescript": "^5.1.6"
+    "typescript": "5.2.2"
   }
 }
