+++ packages/lib/universal/upload/update-file.ts
@@ -3,8 +3,6 @@ import { match } from 'ts-pattern';
 
 import { DocumentDataType } from '@documenso/prisma/client';
 
-import { getAbsolutePresignPostUrl } from './server-actions';
-
 export type UpdateFileOptions = {
   type: DocumentDataType;
   oldData: string;
@@ -40,6 +38,8 @@ const updateFileWithBytes64 = (data: string) => {
 };
 
 const updateFileWithS3 = async (key: string, data: string) => {
+  const { getAbsolutePresignPostUrl } = await import('./server-actions');
+
   const { url } = await getAbsolutePresignPostUrl(key);
 
   const response = await fetch(url, {
