+++ packages/prisma/tsconfig.seed.json
@@ -1,0 +1,6 @@
+{
+  "$schema": "https://json.schemastore.org/tsconfig",
+  "compilerOptions": {
+    "module": "NodeNext"
+  }
+}
