+++ packages/ui/package.json
@@ -22,7 +22,7 @@
     "@types/react": "18.2.18",
     "@types/react-dom": "18.2.7",
     "react": "18.2.0",
-    "typescript": "^5.1.6"
+    "typescript": "5.2.2"
   },
   "dependencies": {
     "@documenso/lib": "*",
@@ -60,11 +60,11 @@
     "framer-motion": "^10.12.8",
     "lucide-react": "^0.279.0",
     "luxon": "^3.4.2",
-    "next": "13.4.19",
-    "pdfjs-dist": "3.6.172",
+    "next": "13.5.4",
+    "pdfjs-dist": "3.11.174",
     "react-day-picker": "^8.7.1",
     "react-hook-form": "^7.45.4",
-    "react-pdf": "^7.3.3",
+    "react-pdf": "^7.5.0",
     "react-rnd": "^10.4.1",
     "tailwind-merge": "^1.12.0",
     "tailwindcss-animate": "^1.0.5"
